pyproject.toml
[project]
name = "fly_in"
version = "0.0.1"
description = "tlaranje's fly-in project (school 42)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flake8>=7.1.1",
    "mypy>=1.11.2",
    "pydantic>=2.8.2",
    "rich>=13.8.1",
    "pygame-ce<=2.5.7"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"